Move tests to dedicated DiscoverySelectorsTests class
